# api/openapi.yaml
openapi: 3.1.0
info:
  title: Customer Portfolio API
  version: 1.0.0
  description: >
    Presentation-layer API for a fictional financial institution. It provides
    deterministic synthetic data for demonstration purposes only.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/customers/{customerId}/portfolio:
    get:
      operationId: getCustomerPortfolio
      summary: Get an aggregated customer portfolio
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            pattern: ^cust-[0-9]{4,12}$
          example: cust-1001
      responses:
        "200":
          description: Aggregated portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamFailure"
...
components:
  schemas:
    Money:
      type: object
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: ^-?[0-9]+\.[0-9]{2}$
          example: "34375.00"
        currency:
          type: string
          pattern: ^[A-Z]{3}$
          example: USD
    Holding:
      type: object
      additionalProperties: false
      required: [symbol, quantity, unitPrice, marketValue]
      properties:
        symbol:
          type: string
          example: AAPL
        quantity:
          type: string
          example: "100"
        unitPrice:
          $ref: "#/components/schemas/Money"
        marketValue:
          $ref: "#/components/schemas/Money"
    Account:
      type: object
      additionalProperties: false
      required: [accountId, accountType, cashBalance, marketValue, holdings]
      properties:
        accountId:
          type: string
          example: acct-brokerage-01
        accountType:
          type: string
          enum: [BROKERAGE, RETIREMENT]
        cashBalance:
          $ref: "#/components/schemas/Money"
        marketValue:
          $ref: "#/components/schemas/Money"
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/Holding"
    Portfolio:
      type: object
      additionalProperties: false
      required: [customerId, asOf, totalMarketValue, accounts]
      properties:
        customerId:
          type: string
          example: cust-1001
        asOf:
          type: string
          format: date-time
          example: "2026-09-05T12:00:00Z"
        totalMarketValue:
          $ref: "#/components/schemas/Money"
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Invalid customer identifier
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Customer was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UpstreamFailure:
      description: A required upstream could not provide valid data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"